Fix problem with select in schema of schema with unnamespaced keys
Within conform for (select), when recursing to a nested schema spec,
the lookup function was constructing a sub-schema form that disconsidered
he structure of the nested schema, resulting in exceptions when
that nested schema was more complex than a simple vector of keywords,
in particular when the schema form corresponted to a map of unnamespaced
keywords.

This attempted fix, grabs the schema form from the schema-obj through
(s/form).