Fix bug in select spec conform
Whenever a select validated a map containing a collection value specificed
with s/coll and the element spec for the s/coll was not a valid schema
spec, conform would throw an IllegalArgumentException.

The fix tries to transform the s/coll element spec into a schema through
s/schema* and if the result is not a schema, it does not recurse into a
subselect, but rather validates with the s/coll spec directly.